+++ src/Controller/Admin/DashboardController.php
@@ -37,7 +37,7 @@ class DashboardController extends AbstractDashboardController
     public function configureDashboard(): Dashboard
     {
         return Dashboard::new()
-            ->setTitle('<img src="/img/itk-sites-logo.png" width="170px" alt="ITK sites">')
+            ->setTitle('<img src="/img/itk-sites-logo.png" width="170px" alt="ITK sites logo">')
             ->setFaviconPath('img/favicon.ico')
             ->renderContentMaximized();
     }
@@ -52,6 +52,7 @@ class DashboardController extends AbstractDashboardController
         yield MenuItem::linkTo(DomainCrudController::class, 'Domains', 'fas fa-link');
         yield MenuItem::linkTo(OIDCCrudController::class, 'OIDC', 'fas fa-key');
         yield MenuItem::linkTo(ServiceCertificateCrudController::class, 'Service certificates', 'fas fa-lock');
+        yield MenuItem::linkTo(SecurityContractCrudController::class, 'Service Agreements', 'fas fa-file-contract');
         yield MenuItem::section('Dependencies');
         yield MenuItem::linkTo(PackageCrudController::class, 'Packages', 'fas fa-cube');
         yield MenuItem::linkTo(PackageVersionCrudController::class, 'Package Versions', 'fas fa-cubes');